Add files via upload
This script creates a social network edge set for the House of Representatives. The end result is a dataset which lists a pair of House members and their shared number of votes.  The script uses rvest to scrape voting data from the Office of the Clerk, U.S. House of Representatives.